Move general query functions to common
