+++ splice/queries/adgroup.py
@@ -3,7 +3,7 @@ from sqlalchemy.sql import exists
 from sqlalchemy.orm.exc import NoResultFound
 from sqlalchemy.exc import InvalidRequestError
 
-from splice.queries import row_to_dict
+from splice.queries.common import row_to_dict
 
 
 def get_adgroups_by_campaign_id(campaign_id):
